msm: ipa3: nullify FW IRAM before loading
The change takes into account the p_memsz and p_filesz
fields in the ELF program header.
p_memsz indicates the entire size of the IRAM, while
p_filesz indicates the actual size of the FW to be loaded.
As part of the loading procedure, the loader is required
to set the entire IRAM region to 0s before loading the FW.

Change-Id: I2ca1258903d9cbfb874a2cee38414f283055c7d9
CRs-Fixed: 970969
Acked-by: David Arinzon <darinzon@qti.qualcomm.com>
Signed-off-by: Sivan Reinstein <sivanr@codeaurora.org>