USB: dwc3-msm: Don't configure DBM for GSI endpoints
On some platforms, USB endpoints are configured as GSI endpoints.
In this case, it is not required to configure USB endpoint as DBM endpoint.
Otherwise it might cause USB functionality failure. Hence add a proper
check so that skip configuring USB endpoints as DBM endpoints in case
of GSI endpoints.

Change-Id: I8548744b816a5e2b8807d1d633743c270deedb8d
Signed-off-by: Vijayavardhan Vennapusa <vvreddy@codeaurora.org>