msm: ipa: use timeout mechanism for holb monitoring command
When we use polling mechanism for holb monitoring command
Uc will still try to send a interrupt to notify the response.
This can result in a sync mismatch between AP and Uc and will
result in a panic. Make changes to use timeout mechanism
for holb monitoring command as well.

Change-Id: I8c74cfedff30f90dceec7f89b153018c307afa0a
Acked-by: Chaitanya Pratapa <cpratapa@qti.qualcomm.com>
Signed-off-by: Ravinder Konka <rkonka@codeaurora.org>